Rename CommandDetail to CommandDetailTable for clarity
