+++ packages/react-ui/src/Tables/CommandDetailTable.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import { CommandDetail, CommandDetailProps } from './CommandDetail'
+import { CommandDetailTable, CommandDetailProps } from './CommandDetailTable'
 
 const props: CommandDetailProps = {
   commands: [
@@ -20,23 +20,23 @@ const props: CommandDetailProps = {
 }
 
 test('should render the component', async () => {
-  expect(() => render(<CommandDetail {...props} />)).not.toThrow()
+  expect(() => render(<CommandDetailTable {...props} />)).not.toThrow()
 })
 
 test('should display the command name', async () => {
-  render(<CommandDetail {...props} />)
+  render(<CommandDetailTable {...props} />)
 
   expect(screen.getByText(props.commands[0].name)).toBeInTheDocument()
 })
 
 test('should display the command description', async () => {
-  render(<CommandDetail {...props} />)
+  render(<CommandDetailTable {...props} />)
 
   expect(screen.getByText(props.commands[0].description)).toBeInTheDocument()
 })
 
 test('should display the pre-populated value in the dropdown if provided', async () => {
-  render(<CommandDetail {...props} />)
+  render(<CommandDetailTable {...props} />)
 
   expect(
     screen.queryByDisplayValue(`${props.commands[0].value}`)
@@ -44,7 +44,7 @@ test('should display the pre-populated value in the dropdown if provided', async
 })
 
 test('should display a checkbox if no parameter options are provided', async () => {
-  render(<CommandDetail {...props} />)
+  render(<CommandDetailTable {...props} />)
 
   expect(screen.queryByRole('checkbox')).toBeInTheDocument()
 })
